OCR/requirements.txt
streamlit>=1.33,<2.0
PyMuPDF>=1.24,<2.0
requests>=2.31,<3.0
numpy>=1.26,<3.0
opencv-python>=4.9,<5.0
Pillow>=10.2,<12.0
easyocr>=1.7,<2.0
torch>=2.2
torchvision>=0.17
